Rewrited datasync, added command manager, added Cyclic actions + some fixes

+++ BotAnbotip/Bot/Commands/RatingListCommands.cs
@@ -6,6 +6,8 @@ using BotAnbotip.Bot.Data;
 using System;
 using System.Collections.Generic;
 using Discord.Rest;
+using BotAnbotip.Bot.Data.Group;
+using BotAnbotip.Bot.Data.CustomClasses;
 
 namespace BotAnbotip.Bot.Commands
 {
@@ -19,12 +21,13 @@ namespace BotAnbotip.Bot.Commands
 
         public static async Task AddListAsync(SocketMessage message, string argument)
         {
-            await message.DeleteAsync();
-
-            var userRoles = ((IGuildUser) message.Author).RoleIds;
-
-            if (userRoles.Contains((ulong) RoleIds.Основатель) || userRoles.Contains((ulong) RoleIds.Администратор))
+            try
             {
+                await message.DeleteAsync();
+                if (!CommandManager.CheckPermission((IGuildUser)message.Author, RoleIds.Основатель)) return;
+
+                var userRoles = ((IGuildUser)message.Author).RoleIds;
+
                 var strBuf = argument.Split(' ');
                 var listName = argument.Substring((strBuf[0].Length));
                 RatingListType listType = RatingListType.Other;
@@ -41,63 +44,77 @@ namespace BotAnbotip.Bot.Commands
                     }
                 }
 
-                var newRatingChannel = await Info.GroupGuild.CreateTextChannelAsync(listName);
+                var newRatingChannel = await ConstInfo.GroupGuild.CreateTextChannelAsync(listName);
                 await newRatingChannel.ModifyAsync((textChannelProperties) =>
                 {
-                    textChannelProperties.CategoryId = (ulong) CategoryIds.Рейтинговые_Листы;
+                    textChannelProperties.CategoryId = (ulong)CategoryIds.Рейтинговые_Листы;
                 });
 
-                await newRatingChannel.AddPermissionOverwriteAsync(Info.GroupGuild.GetRole((ulong) RoleIds.Главный_Бот),
+                await newRatingChannel.AddPermissionOverwriteAsync(ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Главный_Бот),
                     OverwritePermissions.AllowAll(newRatingChannel));
 
                 await newRatingChannel.AddPermissionOverwriteAsync(
-                    Info.GroupGuild.GetRole((ulong) RoleIds.Музыкальный_Бот),
+                    ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Музыкальный_Бот),
                     OverwritePermissions.DenyAll(newRatingChannel));
                 await newRatingChannel.AddPermissionOverwriteAsync(
-                    Info.GroupGuild.GetRole((ulong)RoleIds.Чат_Бот),
+                    ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Чат_Бот),
                     OverwritePermissions.DenyAll(newRatingChannel));
                 await newRatingChannel.AddPermissionOverwriteAsync(
-                    Info.GroupGuild.GetRole((ulong)RoleIds._Бот),
+                    ConstInfo.GroupGuild.GetRole((ulong)RoleIds._Бот),
                     OverwritePermissions.DenyAll(newRatingChannel));
 
-                await newRatingChannel.AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
+                await newRatingChannel.AddPermissionOverwriteAsync(
+                    ConstInfo.GroupGuild.EveryoneRole,
                     OverwritePermissions.DenyAll(newRatingChannel));
-                await newRatingChannel.AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                    newRatingChannel.GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                        null, PermValue.Allow, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
 
-                DataManager.ratingChannels.Add(newRatingChannel.Id,
+                await newRatingChannel.AddPermissionOverwriteAsync(ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Активный_Участник),
+                    OverwritePermissions.DenyAll(newRatingChannel).Modify(PermValue.Allow,
+                        null, null, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+                await newRatingChannel.AddPermissionOverwriteAsync(ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Модератор),
+                    OverwritePermissions.DenyAll(newRatingChannel).Modify(PermValue.Allow,
+                        null, null, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+                await newRatingChannel.AddPermissionOverwriteAsync(ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Администратор),
+                    OverwritePermissions.DenyAll(newRatingChannel).Modify(PermValue.Allow,
+                        null, null, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+                await newRatingChannel.AddPermissionOverwriteAsync(ConstInfo.GroupGuild.GetRole((ulong)RoleIds.Заместитель),
+                    OverwritePermissions.DenyAll(newRatingChannel).Modify(PermValue.Allow,
+                        null, null, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+
+                DataManager.RatingChannels.Add(newRatingChannel.Id,
                     new RatingList(newRatingChannel.Id, newRatingChannel.Name, listType));
-            }
 
-            await DataManager.SaveDataAsync();
+                await DataManager.SaveDataAsync(DataManager.RatingChannels, nameof(DataManager.RatingChannels));
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
         }
 
         public static async Task RemoveListAsync(SocketMessage message, string argument)
         {
             await message.DeleteAsync();
+            if (!CommandManager.CheckPermission((IGuildUser)message.Author, RoleIds.Основатель)) return;
 
             var userRoles = ((IGuildUser) message.Author).RoleIds;
 
             if (userRoles.Contains((ulong) RoleIds.Основатель))
             {
                 ulong id = ulong.Parse(argument);
-                var ratingChannel = Info.GroupGuild.GetChannel(id);
+                var ratingChannel = ConstInfo.GroupGuild.GetChannel(id);
                 if (ratingChannel != null) await ratingChannel.DeleteAsync();
 
                 DataManager.RemoveRatingList(id);
             }
 
-            await DataManager.SaveDataAsync();
+            await DataManager.SaveDataAsync(DataManager.RatingChannels, nameof(DataManager.RatingChannels));
         }
 
         public static async Task AddValueAsync(SocketMessage message, string argument)
         {
-            await ((IGuildChannel)message.Channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                       ((IGuildChannel)message.Channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                           null, PermValue.Deny, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
-
             await message.DeleteAsync();
+            if (!CommandManager.CheckPermission((IGuildUser)message.Author, RoleIds.Основатель)) return;
+
             string[] bufStr = argument.Split(' ');
             const int numOfSpaces = 3;
             bool flag = bufStr[0] == "к+с";
@@ -123,7 +140,7 @@ namespace BotAnbotip.Bot.Commands
             await sendedMessage.AddReactionAsync(new Emoji("💙"));
             await sendedMessage.AddReactionAsync(new Emoji("❌"));
 
-            var ratingList = DataManager.ratingChannels[message.Channel.Id];
+            var ratingList = DataManager.RatingChannels[message.Channel.Id];
             if (flag) ratingList.ListObjects.Add(objName, sendedMessage.Id, url, thumbnailUrl);
             else ratingList.ListObjects.Add(objName, sendedMessage.Id);
 
@@ -132,38 +149,34 @@ namespace BotAnbotip.Bot.Commands
                 await sendedMessage.AddReactionAsync(new Emoji(TypeEmodji[ratingList.Type]));
             }
 
-            await SortAsync(message.Channel, DataManager.ratingChannels[message.Channel.Id].ListObjects[objName]); //не эффективно по аргументу
-            await DataManager.SaveDataAsync();
-
-            await ((IGuildChannel)message.Channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                       ((IGuildChannel)message.Channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                           null, PermValue.Allow, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+            await SortAsync(message.Channel, DataManager.RatingChannels[message.Channel.Id].ListObjects[objName]); //не эффективно по аргументу
+            await DataManager.SaveDataAsync(DataManager.RatingChannels, nameof(DataManager.RatingChannels));
         }
 
         public static async Task RemoveValueAsync(SocketMessage message, string argument)
         {
             await message.DeleteAsync();
-            if (DataManager.ratingChannels[message.Channel.Id].ListObjects[argument] != null)
+            if (!CommandManager.CheckPermission((IGuildUser)message.Author, RoleIds.Основатель)) return;
+
+            if (DataManager.RatingChannels[message.Channel.Id].ListObjects[argument] != null)
             {
                 var foundedMessage =
-                    await message.Channel.GetMessageAsync(DataManager.ratingChannels[message.Channel.Id]
+                    await message.Channel.GetMessageAsync(DataManager.RatingChannels[message.Channel.Id]
                         .ListObjects[argument].MessageId);
                 await foundedMessage.DeleteAsync();
 
-                DataManager.ratingChannels[message.Channel.Id].ListObjects.Remove(argument);
-                await DataManager.SaveDataAsync();
+                DataManager.RatingChannels[message.Channel.Id].ListObjects.Remove(argument);
+                await DataManager.SaveDataAsync(DataManager.RatingChannels, nameof(DataManager.RatingChannels));
             }
         }
 
         public static async Task ChangeRatingAsync(IUserMessage message, ISocketMessageChannel channel, SocketReaction reaction)
         {
-            await ((IGuildChannel)channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                    ((IGuildChannel)channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                        null, PermValue.Deny, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+            if (!CommandManager.CheckPermission((IGuildUser)reaction.User.Value, RoleIds.Активный_Участник)) return;
 
             var objName = ConvertMessageToRatingListObject(message);
 
-            var likedObject = DataManager.ratingChannels[channel.Id].ListObjects[objName];
+            var likedObject = DataManager.RatingChannels[channel.Id].ListObjects[objName];
             Evaluation eval = reaction.Emote.Name == "💙" ? Evaluation.Like : Evaluation.Dislike;
 
             //Проверка на наличие пользователя в массиве и сравнение его оценки
@@ -181,23 +194,15 @@ namespace BotAnbotip.Bot.Commands
                     messageProperties.Embed = embedBuilder.Build();
                 });
                 await SortAsync(channel, likedObject);
-                await DataManager.SaveDataAsync();
-
-                await ((IGuildChannel)channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                    ((IGuildChannel)channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                        null, PermValue.Allow, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
+                await DataManager.SaveDataAsync(DataManager.RatingChannels, nameof(DataManager.RatingChannels));
             }
         }
 
-        private static async Task SortAsync(ISocketMessageChannel channel, ObjectOfRatingList obj)
+        private static async Task SortAsync(ISocketMessageChannel channel, RatingListObject obj)
         {
-            await ((IGuildChannel)channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                    ((IGuildChannel)channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                        null, PermValue.Deny, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
-
             var bufMessage1 = await channel.GetMessageAsync(obj.MessageId);
 
-            DataManager.ratingChannels[channel.Id].ListObjects.Sort(obj);
+            DataManager.RatingChannels[channel.Id].ListObjects.Sort(obj);
 
 
             if (obj.CurrentPosition != obj.PreviousPosition)
@@ -206,12 +211,12 @@ namespace BotAnbotip.Bot.Commands
 
                 for (int i = obj.PreviousPosition; i != obj.CurrentPosition; i -= eval)
                 {
-                    var firstObject = DataManager.ratingChannels[channel.Id].ListObjects[i];
+                    var firstObject = DataManager.RatingChannels[channel.Id].ListObjects[i];
                     var firstMessage =
                         await channel.GetMessageAsync(firstObject.MessageId);
                     await Task.Delay(300);
 
-                    var secondObject = DataManager.ratingChannels[channel.Id].ListObjects[i - eval];
+                    var secondObject = DataManager.RatingChannels[channel.Id].ListObjects[i - eval];
                     var secondMessage =
                         await channel.GetMessageAsync(secondObject.MessageId);
                     await Task.Delay(300);
@@ -220,7 +225,7 @@ namespace BotAnbotip.Bot.Commands
 
                 }
 
-                var bufMessage2 = await channel.GetMessageAsync(DataManager.ratingChannels[channel.Id]
+                var bufMessage2 = await channel.GetMessageAsync(DataManager.RatingChannels[channel.Id]
                     .ListObjects[obj.CurrentPosition].MessageId);
                 await ((IUserMessage) bufMessage2).ModifyAsync((messageProperties) =>
                 {
@@ -233,15 +238,10 @@ namespace BotAnbotip.Bot.Commands
 
                     messageProperties.Embed = embedBuilder.Build();
                 });
-
             }
-
-            await ((IGuildChannel)channel).AddPermissionOverwriteAsync(Info.GroupGuild.EveryoneRole,
-                    ((IGuildChannel)channel).GetPermissionOverwrite(Info.GroupGuild.EveryoneRole).Value.Modify(PermValue.Allow,
-                        null, PermValue.Allow, PermValue.Allow, null, null, null, null, null, PermValue.Allow));
         }
 
-        private static async Task SwapTwoMessage(IMessage firstMessage, ObjectOfRatingList firstObject, IMessage secondMessage, ObjectOfRatingList secondObject)
+        private static async Task SwapTwoMessage(IMessage firstMessage, RatingListObject firstObject, IMessage secondMessage, RatingListObject secondObject)
         {
             await ((IUserMessage)firstMessage).ModifyAsync((messageProperties) =>
             {
